Create No twins plz
No twins pls !!!: Mr. Chitti, a robot is assigned a job of reporting to his boss Dr. Vaseegaran
about the existence of twins in a group. Chitti’s job is just to report whether the group is distinct
or not. Even if one duplicate is present then the group is not unique. Help Mr. Chitti to report.
Input: an array of integers
output: Array distinct / Array not distinct
Example: Consider the following array. The array is not distinct as 10 is duplicated.